Latest testing of zc_graph_.

+++ include/displays/s11_rx.hpp
@@ -42,7 +42,7 @@ namespace display_modes {
 			params_.title = "S11 Resistance and Reactance vs frequency";
 
 			axis_params_t x_axis_params;
-			x_axis_params.outer_range = { 0.0F, FLT_MAX };
+			x_axis_params.outer_range = { 0.0F, DBL_MAX };
 			x_axis_params.inner_range = { 1e6F, 30e6F };
 			x_axis_params.default_range = { 1e6F, 30e6F };
 			x_axis_params.unit_modifier = zc_graph_::modifier_t::SI_PREFIX;
@@ -75,7 +75,7 @@ namespace display_modes {
 
 		void convert_sp_point(
 			const sp_point& point,
-			float Z0,
+			double Z0,
 			zc_graph_::data_point_t& point_l,
 			zc_graph_::data_point_t& point_r) const
 		{
